+++ server/api/src/main/java/com/walmartlabs/concord/server/api/ServerResource.java
@@ -12,4 +12,9 @@ public interface ServerResource {
     @Path("/ping")
     @Produces(MediaType.APPLICATION_JSON)
     PingResponse ping();
+
+    @GET
+    @Path("version")
+    @Produces(MediaType.APPLICATION_JSON)
+    VersionResponse version();
 }